clean up code; add xml docs
