+++ JsonPath/Expressions/BinaryValueExpressionNode.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Text;
+﻿using System.Text;
 using System.Text.Json.Nodes;
 
 namespace Json.Path.Expressions;
