+++ JsonPath/PathEvaluationOptions.cs
@@ -1,6 +1,12 @@
 ﻿namespace Json.Path;
 
+/// <summary>
+/// Provides a number of options to be used during evaluation.
+/// </summary>
+/// <remarks>
+/// Currently that number is zero.
+/// </remarks>
 public class PathEvaluationOptions
 {
-	public bool Strict { get; set; } // TODO
+//	public bool Strict { get; set; } // TODO
 }
